Add experiment's model

+++ src/main/java/com/example/spring/demo/service/ExperimentService.java
@@ -1,9 +1,12 @@
 package com.example.spring.demo.service;
 
 import com.example.spring.demo.entity.ExperimentEntity;
+import com.example.spring.demo.model.Experiment;
 import com.example.spring.demo.repository.ExperimentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
 
 @Service
 public class ExperimentService {
@@ -13,34 +16,39 @@ public class ExperimentService {
         this.experimentRepository = experimentRepository;
     }
 
-    public ExperimentEntity registration(ExperimentEntity experiment) {
-        return experimentRepository.save(experiment);
+    public Experiment registration(ExperimentEntity experiment) {
+        return Experiment.toModel(experimentRepository.save(experiment));
     }
 
-    public Iterable<ExperimentEntity> findAll() {
-        return experimentRepository.findAll();
+    public ArrayList<Experiment> findAll() {
+        ArrayList<Experiment> experiments = new ArrayList<>();
+
+        for (ExperimentEntity entity : experimentRepository.findAll()) {
+            experiments.add(Experiment.toModel(entity));
+        }
+
+        return experiments;
     }
 
-    public ExperimentEntity findById(Integer id) {
-        return experimentRepository.findById(id).get();
+    public Experiment findById(Integer id) {
+        return Experiment.toModel(experimentRepository.findById(id).get());
     }
 
-    public ExperimentEntity findByName(String name) {
-        return experimentRepository.findByName(name);
+    public Experiment findByName(String name) {
+        return Experiment.toModel(experimentRepository.findByName(name));
     }
 
     public Integer deleteById(Integer id) {
-        experimentRepository.deleteById(id);
-        return id;
+        if (experimentRepository.existsById(id)) {
+            experimentRepository.deleteById(id);
+            return 0;
+        }
+        return -1;
     }
 
     public Integer deleteAll() {
         experimentRepository.deleteAll();
         return 0;
-    }
-
-    public boolean existById(Integer id) {
-        return experimentRepository.existsById(id);
     }
 
 }
